assignment operators.
Sidenote: I am using Jim Hester's lintr and emacs' flycheck-mode to pick
up most of these style incongruities.  However, I am trying to change
only one class of disagreement at a time, thus I likely am missing some.

In this pass of the code, I attempted to make the assignments more
consistent.  R provides 'a = b' and 'a <- b' as you know.
I always thought having two characters to assign values a bit stupid.
But, R provides 'a <<- b' to push the assignment up to a parent
environment, this seems like a logical reason to choose 'a <- b'; I
understand though if one doesn't like the arrow styles his/her code
with ' = '.

With that in mind, having a mix of 'a  <-  b', 'a<-b', 'a  =  b', 'a=b'
and every combination in between makes it look like the author does not
give a crap about what he/she is doing.

In addition, I attempted to remove any instances of:
'a=b;c=d;e=f'.  I could sort of understand this if one were really
trying to keep the codebase as concise as possible; there are well over
400 (to my eyes) spurious blank lines scattered through the code, thus
there seems no logical reason to do this.